Combine boilerplate and reactNavigation templates
