+++ templates/boilerplate/src/navigators/DashboardStack.tsx
