+++ templates/boilerplate/src/navigators/TabNavigator.tsx
