+++ templates/boilerplate/src/screens/InformationScreen/InformationScreen.tsx
