--- templates/reactNavigation/src/navigators/navigatorTypes.tsx
@@ -1,6 +1,0 @@
-// add types for navigation here
-// key: screen name
-// value: params (use undefined if accepts none)
-export type RootStackParamList = {
-  Home: undefined;
-};
