--- templates/reactNavigationBottomTabs/src/types/global.d.ts
@@ -1,9 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { RootStackParamList } from '../navigators/navigatorTypes';
-
-declare global {
-  namespace ReactNavigation {
-    // eslint-disable-next-line @typescript-eslint/no-empty-interface
-    interface RootParamList extends RootStackParamList {}
-  }
-}
